Replaced Profiles Class with Users class
This replaces Profiles.class as it reads better.

Added several Methods:

user_selection - Which Checks for and Creates a folder called Profiles

user_list - Checks to see if there is a file within the Profiles folder if true prints a list of users.

no_creation_of_profile - If a user doesn't create a profile then asks again if they want to else it quits